refactor multi accounts component
