+++ src/components/multiAccounts/activeAccount.tsx
@@ -1,25 +1,23 @@
 import { RelayContext } from '@components/relaysProvider';
 
-import { dateToUnix } from '@utils/getDate';
+import { DEFAULT_AVATAR } from '@stores/constants';
+
 import { createFollows } from '@utils/storage';
 import { tagsToArray } from '@utils/transform';
 
 import * as DropdownMenu from '@radix-ui/react-dropdown-menu';
 import { AvatarIcon, ExitIcon, GearIcon } from '@radix-ui/react-icons';
 import { writeText } from '@tauri-apps/api/clipboard';
-import destr from 'destr';
 import Image from 'next/image';
 import { useRouter } from 'next/router';
 import { nip19 } from 'nostr-tools';
-import { memo, useContext, useEffect, useRef } from 'react';
+import { memo, useContext, useEffect } from 'react';
 
 export const ActiveAccount = memo(function ActiveAccount({ user }: { user: any }) {
   const [pool, relays]: any = useContext(RelayContext);
 
   const router = useRouter();
-  const userData = destr(user.metadata);
-
-  const now = useRef(new Date());
+  const userData = JSON.parse(user.metadata);
 
   const openProfilePage = () => {
     router.push(`/users/${user.id}`);
@@ -35,7 +33,6 @@ export const ActiveAccount = memo(function ActiveAccount({ user }: { user: any }
         {
           kinds: [3],
           authors: [user.id],
-          since: dateToUnix(now.current),
         },
       ],
       relays,
@@ -61,12 +58,10 @@ export const ActiveAccount = memo(function ActiveAccount({ user }: { user: any }
       <DropdownMenu.Trigger asChild>
         <button className="relative h-11 w-11 rounded-md">
           <Image
-            src={userData.picture}
+            src={userData.picture || DEFAULT_AVATAR}
             alt="user's avatar"
             fill={true}
             className="rounded-md object-cover"
-            placeholder="blur"
-            blurDataURL="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNkqAcAAIUAgUW0RjgAAAAASUVORK5CYII="
             priority
           />
         </button>
