+++ src/stores/constants.tsx
@@ -1,1 +1,2 @@
+export const APP_VERSION = '0.2.1';
 export const DEFAULT_AVATAR = 'https://void.cat/d/KmypFh2fBdYCEvyJrPiN89.webp';
